cargo-make: adopted cargo-make as the project build tool (#16)
Adopted cargo-make as the project build tool.

Provided targets for building, testing, linting, documenting, and performing
project-wide integration tests.

Fixed various issues stopping documentation from being generated.

Fixed various issues dredged up by cargo-clippy.

Updated README.markdown.

+++ libsupervisionary/src/raw/theorem.rs
@@ -10,8 +10,7 @@
 //! `LICENSE.markdown` file in the *Supervisionary* root directory for licensing
 //! information.
 //!
-//! [Dominic Mulligan]: https://dominicpm.github.io
-//! [Arm Research]: http://www.arm.com/research
+//! [Dominic Mulligan]<https://dominicpm.github.io>
 
 use crate::raw::{tags, ErrorCode, Handle, Name, RawHandle};
 use std::{convert::TryFrom, marker::PhantomData};
@@ -928,7 +927,7 @@ where
 
     let status = unsafe {
         __theorem_register_forall_introduction(
-            name_handle.into() as u64,
+            name_handle.into(),
             *type_handle.into() as u64,
             *theorem_handle.into() as u64,
             &mut result as *mut u64,
